Removed extra subshell to fix zombie processes
The multi-command handling is now

    bash -c " ... "

instead of

    bash -c "( ... )"

Previously the extra ( ) woudl start a subshell, which would prevent us from stopping the python process.
(kill signal was going to shell, but python remained as zombie process)

Also ran into a weird bash behavior hat is different on vader (older bash) and cloud-kitchen (newer bash).

Consider the command:

    nohup bash -c 'echo "helloz" && python -c "import time; time.sleep(10)"' & echo $!

On cloud kitchen, the $! expands to the python process.
On vader, this the $! expands to the bash process that called python:

The stop_daemon_chef now handles both cases...  there goes a day I'm not
gettin back! But hey I learned a lot about bash, subshells, etc.